Move schema builder functions to utils
